+++ abrg/internal/util/char.go
@@ -2,10 +2,10 @@ package util
 
 import "abrg/internal/char"
 
-// isKanjiNumeral reports whether r is a kanji numeral (一二三四五六七八九十〇百千).
+// IsKanjiNumeral reports whether r is a kanji numeral (一二三四五六七八九十〇百千).
 // The set deliberately differs from transform's kanji-number set: 零/万/億
 // never appear in address numbers, so they are excluded here.
-func isKanjiNumeral(r rune) bool {
+func IsKanjiNumeral(r rune) bool {
 	switch r {
 	case '一', '二', '三', '四', '五', '六', '七', '八', '九', '十', '〇', '百', '千':
 		return true
@@ -19,5 +19,5 @@ func IsAddressNumberRune(r rune) bool {
 	if char.IsASCIIDigit(r) || r >= '０' && r <= '９' {
 		return true
 	}
-	return isKanjiNumeral(r)
+	return IsKanjiNumeral(r)
 }
